+++ lib/src/esc/esc.dart
@@ -1,4 +1,4 @@
-import 'common.dart';
+import '../control_codes/control_codes.dart';
 
 /// Saves the cursor.
 ///
@@ -11,9 +11,3 @@ const String saveCursor = '${esc}7';
 /// attributes from the previous [saveCursor] if any, otherwise resets these
 /// all to their defaults.
 const String restoreCursor = '${esc}8';
-
-/// Link.
-///
-/// https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
-String link(String url, [String? text]) =>
-    '$esc]8;;$url$bel${text ?? url}$esc]8;;$bel';